+++ temba/orgs/models.py
@@ -292,6 +292,10 @@ class Org(SmartModel):
         null=True, max_length=128, default=None, help_text=_("A password that allows users to register as surveyors")
     )
 
+    flow_server_enabled = models.BooleanField(
+        default=False, help_text=_("Whether flows and messages should be handled by the flow server")
+    )
+
     parent = models.ForeignKey(
         "orgs.Org",
         on_delete=models.PROTECT,
@@ -2439,7 +2443,7 @@ class Language(SmartModel):
         return dict(name=self.name, iso_code=self.iso_code)
 
     @classmethod
-    def get_localized_text(cls, text_translations, preferred_languages, default_text=None):
+    def get_localized_text(cls, text_translations, preferred_languages, default_text=""):
         """
         Returns the appropriate translation to use.
         :param text_translations: A dictionary (or plain text) which contains our message indexed by language iso code