+++ temba/orgs/views.py
@@ -1169,7 +1169,7 @@ class OrgCRUDL(SmartCRUDL):
             return "%s %s - %s" % (obj.created_by.first_name, obj.created_by.last_name, obj.created_by.email)
 
     class Update(SmartUpdateView):
-        fields = ("name", "brand", "parent", "is_anon")
+        fields = ("name", "brand", "parent", "is_anon", "flow_server_enabled")
 
         class OrgUpdateForm(forms.ModelForm):
             parent = forms.IntegerField(required=False)
@@ -1181,7 +1181,16 @@ class OrgCRUDL(SmartCRUDL):
 
             class Meta:
                 model = Org
-                fields = ("name", "slug", "stripe_customer", "is_active", "is_anon", "brand", "parent")
+                fields = (
+                    "name",
+                    "slug",
+                    "stripe_customer",
+                    "is_active",
+                    "is_anon",
+                    "flow_server_enabled",
+                    "brand",
+                    "parent",
+                )
 
         form_class = OrgUpdateForm
 
@@ -1248,6 +1257,11 @@ class OrgCRUDL(SmartCRUDL):
                     self.get_object().release()
                 return HttpResponseRedirect(self.get_success_url())
             return super().post(request, *args, **kwargs)
+
+        def post_save(self, obj):
+            # make sure all our flows have flow server enabled according to the org settings
+            if obj.flow_server_enabled:
+                obj.flows.update(flow_server_enabled=True)
 
     class Accounts(InferOrgMixin, OrgPermsMixin, SmartUpdateView):
         class PasswordForm(forms.ModelForm):